modified:   clientSide/package-lock.json 	modified:   clientSide/package.json 	modified:   clientSide/public/index.html 	modified:   clientSide/src/App.jsx 	modified:   clientSide/src/actions/analysis.js 	modified:   clientSide/src/api/index.js 	modified:   clientSide/src/components/PickingsAnalysis.jsx 	modified:   clientSide/src/pages/Dashboard.jsx 	modified:   clientSide/src/pages/ExpenseDetails.jsx 	modified:   clientSide/src/pages/PickingDetails.jsx 	modified:   server/controllers/picking.js 	modified:   server/logs/reqLogs.txt 	modified:   server/middleware/auth.js

+++ clientSide/src/pages/ExpenseDetails.jsx
@@ -19,8 +19,13 @@ const ExpenseDetails = () => {
 
   const handleDelete = (e) => {
     e.preventDefault();
-    dispatch(deleteExpense(id));
-    navigate('/expenses');
+    let confirm = window.confirm('Are you sure you want to delete Expenditure no: '  + id + '?');
+    if (confirm) {
+      dispatch(deleteExpense(id));
+      navigate('/expenses');
+    } else {
+      return;
+    }
   }
   
   const handlePaid = (e) => {
